All callback tests moved to lib.ava.ts
