+++ sputnikdao2/tests-ava/__tests__/lib.ava.ts
@@ -1,5 +1,7 @@
-import { BN, NearAccount, captureError, toYocto, tGas, DEFAULT_FUNCTION_CALL_GAS, Gas, } from 'near-workspaces-ava';
+import { BN, NearAccount, captureError, toYocto, tGas, DEFAULT_FUNCTION_CALL_GAS, Gas, NEAR } from 'near-workspaces-ava';
 import { workspace, initStaking, initTestToken, STORAGE_PER_BYTE, workspaceWithoutInit } from './utils';
+import { voteApprove } from './proposals.ava';
+import { DEADLINE, BOND, proposeBounty, voteOnBounty, claimBounty, doneBounty } from './bounties.ava'
 import * as fs from 'fs';
 
 const DAO_WASM_BYTES: Uint8Array = fs.readFileSync('../res/sputnikdao2.wasm');
@@ -181,4 +183,118 @@ workspace.test('Remove blob', async (test, { root, dao, alice }) => {
     const rootAmountAfterRemove = (await root.balance()).total
     test.false(await dao.view('has_blob', { hash: hash }));
     test.assert(rootAmountAfterRemove.gt(rootAmountBeforeRemove));
+});
+
+workspace.test('Callback for BountyDone', async (test, { alice, root, dao }) => {
+    //During the callback the number bounty_claims_count should decrease
+    const proposalId = await proposeBounty(alice, dao);
+    await voteOnBounty(root, dao, proposalId);
+    await claimBounty(alice, dao, proposalId);
+    await doneBounty(alice, alice, dao, proposalId);
+    //Before the bounty is done there is 1 claim
+    test.is(await dao.view('get_bounty_number_of_claims', {id: 0}), 1);
+    const balanceBefore: NEAR = (await alice.balance()).total;
+    //During the callback this number is decreased
+    await voteOnBounty(root, dao, proposalId + 1);
+    const balanceAfter: NEAR = (await alice.balance()).total;
+    test.is(await dao.view('get_bounty_number_of_claims', {id: 0}), 0);
+    test.assert(balanceBefore.lt(balanceAfter));
+});
+
+workspace.test('Callback transfer', async (test, { alice, root, dao }) => {
+    const user1 = await root.createAccount('user1');
+    // Fail transfer by transfering to non-existent accountId
+    let transferId: number = await user1.call(
+        dao,
+        'add_proposal', {
+        proposal: {
+            description: 'give me tokens',
+            kind: {
+                Transfer: {
+                    token_id: "",
+                    receiver_id: "broken_id",
+                    amount: toYocto('1'),
+                }
+            }
+        },
+    }, { attachedDeposit: toYocto('1') });
+    let user1Balance = (await user1.balance()).total
+    await voteApprove(root, dao, transferId);
+    let { status } = await dao.view('get_proposal', { id: transferId });
+    test.is(status, 'Failed');
+    test.assert((await user1.balance()).total.eq(user1Balance)); // no bond returns on fail
+
+    // now we transfer to real accountId
+    transferId = await user1.call(
+        dao,
+        'add_proposal', {
+        proposal: {
+            description: 'give me tokens',
+            kind: {
+                Transfer: {
+                    token_id: "",
+                    receiver_id: alice.accountId, // valid id this time
+                    amount: toYocto('1'),
+                }
+            }
+        },
+    }, { attachedDeposit: toYocto('1') });
+    user1Balance = (await user1.balance()).total
+    await voteApprove(root, dao, transferId);
+    ({ status } = await dao.view('get_proposal', { id: transferId }));
+    test.is(status, 'Approved');
+    test.assert((await user1.balance()).total.gt(user1Balance)); // returns bond
+});
+
+workspace.test('Callback function call', async (test, { alice, root, dao }) => {
+    const testToken = await initTestToken(root);
+    let transferId: number = await root.call(
+        dao,
+        'add_proposal', {
+        proposal: {
+            description: 'give me tokens',
+            kind: {
+                FunctionCall: {
+                    receiver_id: testToken.accountId,
+                    actions: [{ method_name: 'fail', args: Buffer.from('bad args').toString('base64'), deposit: toYocto('1'), gas: tGas(10) }],
+                }
+            }
+        },
+    }, { attachedDeposit: toYocto('1') });
+    await root.call(dao, 'act_proposal',
+        {
+            id: transferId,
+            action: 'VoteApprove'
+        },
+        {
+            gas: tGas(200),
+        });
+    let { status } = await dao.view('get_proposal', { id: transferId });
+    test.is(status, 'Failed');
+
+    transferId = await root.call(
+        dao,
+        'add_proposal', {
+        proposal: {
+            description: 'give me tokens',
+            kind: {
+                FunctionCall: {
+                    receiver_id: testToken.accountId,
+                    actions: [
+                        { method_name: 'mint', args: Buffer.from('{"account_id": "' + alice.accountId + '", "amount": "10"}').toString('base64'), deposit: '0', gas: tGas(10) },
+                        { method_name: 'burn', args: Buffer.from('{"account_id": "' + alice.accountId + '", "amount": "10"}').toString('base64'), deposit: '0', gas: tGas(10) }],
+                }
+            }
+        },
+    }, { attachedDeposit: toYocto('1') });
+    await root.call(dao, 'act_proposal',
+        {
+            id: transferId,
+            action: 'VoteApprove'
+        },
+        {
+            gas: tGas(200),
+        });
+    ({ status } = await dao.view('get_proposal', { id: transferId }));
+    test.is(status, 'Approved');
 });
